Moved the FAQ container into it's own directory.
